[fx] Add ffx device discovery to `fx shell`
This is a fix-forward of fxr/595456. It includes the changes originally
contained in fxr/591990 and fxr/592608, plus an update to `fx
serve-updates`. The original change didn't work with `fx serve`, because
that script used a call to `get-device-ssh-port` directly, as well as a
call to `fx shell`. The discrepancy between the two results led to the
`fx serve` script looping over the call to `fx shell` repeatedly,
waiting for a connection and discarding the error message it was
returning indicating it couldn't reach the target.

This CL replaces the call in serve-updates with one to the new
`get-fuchsia-device-port`, so the results in `serve-updates` are always
consistent.

Change-Id: If9200470b798ff6b51cea4edfc653d028b0a30f4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/594885
Reviewed-by: Gary Boone <gboone@google.com>
Commit-Queue: Daniel Dougherty <doughertyda@google.com>