[a11y] Modify "scene ready" condition in semantics tests.
The web semantics tests hang during setup waiting for a
ViewStateChanged event indicating that the web view is rendering. Scenic
sends this event on the first render pass that visits the web view after
it's been attached to the scene. It's possible that there's some
non-determinism about when this render pass will happen given that we're
not immediately loading any web content into the view (in which case we
may not request a draw operation). Since we really only care if the
web view is attached to the scene at this point (and not that it's
actually rendering content), we can instead wait until the test fixture
view is attached to the scene and the web view is connected to its
view holder.

TEST: ran flutter and web semantics tests 100 times each.
Bug: 86243
Change-Id: Id12fe60798539d6a3f8e98f92dfdf03ad47c841d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/595385
Reviewed-by: Lucas Radaelli <lucasradaelli@google.com>
Reviewed-by: David Worsham <dworsham@google.com>
Commit-Queue: Alexander Brusher <abrusher@google.com>