[gigaboot] start adding TCP6 protocol
TCP should help flashing speed significantly compared to UDP, for
devices that have the required drivers.

This is a fairly complex EFI API, so we wrap it in a simpler API that
meshes better with our existing fastboot code and hides the parts we
don't care about.

This initial CL just adds the open/close functions, along with related
tests and test fixtures. Follow-on CLs will add functionality for
accepting clients and read/write.

Bug: 68692
Change-Id: I80aaf568efa65353a7799750d3548d6d937a36f9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/594143
Commit-Queue: David Pursell <dpursell@google.com>
Reviewed-by: Anirudh Mathukumilli <rudymathu@google.com>