[fuchsia-async] socket2::Socket to DatagramSocket
Add a method to construct a DatagramSocket from a socket2::Socket. This
is necessary for use cases which need to construct a socket directly
via the fuchsia.posix.socket FIDL APIs. The function checks to ensure
that the socket is of type DGRAM, and sets it to be non-blocking.

Change-Id: I32676e83337411c7fc1b823d731dc924d4d9af6b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/594976
Reviewed-by: Adam Perry <adamperry@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Tony Gong <gongt@google.com>