[blackout] fix tests for local running
The integration between the host and target code of blackout was not
tested, so the mechanisms used to run the target code on a device
rotted. This fixes the rot, now blackout runs again on local devices. A
followup commit will add the necessary things to make it run in CI.

I'm not including a test here for the host-target interaction because
the current strategy is deprecated and fragile, and I'm investigating a
more robust strategy, but I want to get the fixes in because the higher
priority is to get this running again.

Bug: 44392
Change-Id: Icc35f0c8c786463067b0e98b3b77bca38db38dd3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/571550
Reviewed-by: Brandon Castellano <bcastell@google.com>
Commit-Queue: Stephen Demos <sdemos@google.com>