[libc] Add GWP-ASan note in libc.so
A new ELF note ".gwp-asan-info" is added to the libc.so which could be
used by out-of-process crash handlers to locate the GWP-ASan state. The
note contains an relative offset to __libc_gwp_asan_info, which is
initialized to contain 2 pointers to the allocator state and the
allocator metadata during libc startup.

Crash handlers are expected to use the functions and data structures
provided in //zircon/system/ulib/c/scudo:gwp-asan-common.

An integration test for GWP-ASan is also added.

Fixed: 78949
Change-Id: Id19a7f3a8ee608ae2677fdd7866db01e90a6fca9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/593054
Reviewed-by: Francois Rousseau <frousseau@google.com>
Reviewed-by: Carlos Pizano <cpu@google.com>
Reviewed-by: Shai Barack <shayba@google.com>
Commit-Queue: Dangyi Liu <dangyi@google.com>