[brcmfmac] Make all devices have the same parent
Currently brcmfmac creates a root device that is bindable, binding to
the wlanphy driver, then it adds children to that device. This
potentially causes problem for device manager because child devices are
now binding during binding of parent devices.

This change adds the children to the same parent that the first wlanphy
device was added to. In practice this makes the wlanif devices siblings
of the wlanphy device instead of children. This allows the wlanphy
device to safely bind to the wlanphy driver.

Bug: 86567
Test: brcmfac_test, brcmfmac_sim_test, manual testing
Change-Id: I31cf435069a72806b8de3d0c6f2ddd140613571e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/594750
Reviewed-by: Garratt Gallagher <garratt@google.com>
Commit-Queue: Bjoern Johansson <bjoernj@google.com>