[fxfs] (Reland) Periodically flush files
This is a reland of fxrev.dev/565807, with a bug fix: the terminating
future in the flush task's loop was not fused, which results in
undefined behaviour when polling that future in a loop.

Adds a per-volume background flushing task which periodically sweeps
through all files and attempts to flush them.

Attempting to flush every file is slightly inefficient, since we always
take a transaction lock when we attempt flushing. We can make this
better by storing a dirty bit in the CachingObjectHandle (or in the
WritebackCache) so no lock needs to be taken on the object, but this can
be done in a follow-up.

Tested: Manually verified that the flush task periodically runs.
Tested: Unit.
Change-Id: Icda93a89029eebbac80892323934029d630cb13b
Multiply: fxfs-test: 100
Multiply: fxfs-append-tests: 500
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/568806
Commit-Queue: James Sullivan <jfsulliv@google.com>
Reviewed-by: Chris Suter <csuter@google.com>