[starnix] Combine TaskOwner and SyscallContext into CurrentTask
The idea is that CurrentTask is an object only available from the thread
that handles the task's syscalls. Anything thread-local can go in it,
and anything that must run on the task's thread can take it as an
argument.

Change-Id: I69d78eeeb4b5773c27b7382563f0f1c8c889b7b4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/595628
Fuchsia-Auto-Submit: Theodore Dubois <tbodt@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Kevin Lindkvist <lindkvist@google.com>
Reviewed-by: Adam Barth <abarth@google.com>