[assembly] Add crate for PartitionsConfig
This crate allows for the reading/writing of a partitions config, which
is described in src/developer/ffx/plugins/assembly/README.md. This
config lists the partition table of a specific device and the mapping to
system images.

Bug: 84902
Change-Id: Ia793d76558a29e60d61d25b7cf33c1f3c3e9ab76
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/594524
Commit-Queue: Aidan Wolter <awolter@google.com>
Fuchsia-Auto-Submit: Aidan Wolter <awolter@google.com>
Reviewed-by: Aaron Wood <aaronwood@google.com>