[bootserver] Retry downloading images on CRC failures
We've been seeing a fair amount of checksum failures when downloading
images lately, and it seems to be transient (doesn't happen consistently
for a given blob) and not necessarily associated with the Go GCS client,
as other teams have seen similar failures with other GCS clients.

The official GCS docs say it's okay to retry a download in case of a CRC
failure, so this is probably fine to have anyway:
https://cloud.google.com/storage/docs/hashes-etags#validation

Unfortunately the storage library doesn't export the bad CRC error:
https://github.com/googleapis/google-cloud-go/blob/07c804b08b9f2bbe181ffcee1b9b41005463b057/storage/reader.go#L415
So we must resort to a substring check to see if a download error is
caused by a CRC failure.

Also apply retries to corruption errors from the `flate` library, which
are suspected to have a similar root cause.

Bug: 70972
Bug: 84619
Change-Id: I93aef3f2bb1439502c89c961900bb8c550966de2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/595882
Commit-Queue: Oliver Newman <olivernewman@google.com>
Fuchsia-Auto-Submit: Oliver Newman <olivernewman@google.com>
Reviewed-by: Ina Huh <ihuh@google.com>