[audio][intel-hda] Fix race in completer replies
The completers and any member of StreamChannel objects need to be
protected by obj_lock_ given that the stream_channel_ is protected.
Moved the locking to the call forwarding done when StreamChannel
addresses are taken and added a comment to make this clear.

This was uncovered as a flake.

Fixed: 85445

Test: fx test intel-hda-stream-test
Multiply: intel-hda-stream-test

Change-Id: I941eae62f0d16eb75ff9cbd120e1a2a82d484153
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/596066
Reviewed-by: David Greenaway <dgreenaway@google.com>
Commit-Queue: Andres Oportus <andresoportus@google.com>