[usb][xhci] check device state
Originally, we were getting transfer events after a device was detached,
which could happen because ordering of events is not guaranteed. By
accessing the array through the index (GetDeviceState()[i]), we
weren't checking whether or not the device was still connected. Thus
causing a page fault when trying to get the address to the ring.

Now, by accessing the array through the method (GetDeviceState(i)), the
the method checks if the device is connected and returns null if it
isn't. We check if the returned state is null and abort the transfer
event when the device is already disconnected.

Bug:86967

Test: Tested on Atlas

Change-Id: I3d75b14cb8ea1fef21ecf0f64b2d037a88b6f179
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/596061
Reviewed-by: Sam Hansen <hansens@google.com>
Commit-Queue: Ruby Zhuang <rdzhuang@google.com>