[fidl] Remove option bits from fuchsia.fs.GetInfo.
These make it difficult to call and difficult to implement. The
performance of computing the flags is trivial except for the device path
which should itself be relatively quick. We do not forsee adding
additional difficult-to-compute data to this structure.

This is more important as part of bug 84558 where we want
DirectoryAdmin.QueryFilesystem to have the same behavior. That is
called/implemented in more places so these bits become more
inconvenient.

Change-Id: Ied25a9f487398b0e687239b27ad188a2b19ecf02
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/589494
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Reviewed-by: Chris Suter <csuter@google.com>
Commit-Queue: Brett Wilson <brettw@google.com>
API-Review: Chris Suter <csuter@google.com>