[setui] Cancel pending jobs when sources closed
This CL adds cancelation logic to Jobs so that any pending job requests
are removed when there is no longer a way to get back to the client of a
Source. This CL also adds and publishes new events for the cases where a
source is created or ends.

Fixed: 73414
Test: fx test -o setui_service_tests and manually added new fidl to
manually test unsupported behavior
Multiply: setui_service_tests

Change-Id: I9ef36aab8707deb7140ccbc7a4f6eb7f83becbf2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/588710
Reviewed-by: Sijie Chen <sijiec@google.com>
Reviewed-by: William Xiao <wxyz@google.com>
Commit-Queue: Paul Faria <paulfaria@google.com>