[nand] Add a read cache for poorly performing reads.
Adds a small simple FIFO-evicting read cache. The cache should not need
to be very large as it really is intended to re-serve entries until they
can be copied into a newer place for more reliable readback. The small
size allows for a simple implementation without need to optimize
retrieval.

Pages are added to the cache after a successful read that saw either
previously failed reads or reads with >= half of the correctable bits
flipped. Pages are evicted from the cache whenever their block is
erased.

Bug: 85915

Change-Id: Iec7cff8ac6440e9ac5548bf8529f47a9ce536934
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/595904
Commit-Queue: Martin Lindsay <mlindsay@google.com>
Reviewed-by: Chris Suter <csuter@google.com>