[media/vnext] add StreamQueue to vnext/lib
This CL is part of a prototype for media vnext. StreamQueue is a
thread-safe queue whose elements are packets, end-of-stream
indications and clear requests.

StreamQueue is part of vnext/lib/stream_sink, which will also contain
implementations of both ends of fuchsia::media2::StreamSink, to be
added in future CLs.

Test: unit test provided
Change-Id: I8601a191df6b7d22f3e3760a791bf4a53118b723
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/595914
Commit-Queue: Dale Sather <dalesat@google.com>
Reviewed-by: Tom Bergan <tombergan@google.com>
Reviewed-by: Austin Foxley <afoxley@google.com>