[rust][rbe] Include .so with .rlib deps.
Include .so with .rlib (for those crates that have them), to get more
consistent results between local and remote building.
The depfile generation revealed that the .so was missing in some cases,
which caused compilation/linking to fall-back to using the .rlib, when
the .so was actually intended.

https://github.com/rust-lang/rust/issues/90106

Bug: 86896
Change-Id: Ic652496bdae0095cad84127479588b3e3da8df66
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/596429
Commit-Queue: David Fang <fangism@google.com>
Reviewed-by: Tyler Mandry <tmandry@google.com>
Reviewed-by: Jay Zhuang <jayzhuang@google.com>