[ffx snapshot] improve high BDP performance
ffx, Overnet and FIDL provide a good deal of convenience that mostly
allow implementors to ignore most aspects of network side effects. In
the case of latency and total throughput however, latency and
bandwdidth-delay-product concerns remain and must be addressed at the
application protocol layer or within the specific application.

In this case users have observed that particularly in "remote
workflows", that is workflows wherein the device is on the far end of an
SSH tunnel from the ffx daemon, the performance of this command is too
slow. Basic usage of fuchsia.io would only keep 8kb at a time in flow
between round trips - a classic bandwidth-delay-product problem.

This change introduces a solution that operates in two parts:
- Introduce concurrency and asynchrony to requests by issuing requests
  for more data asynchronously from the processing of each response.
- Introduce a batch of asynchrony so as to increase the total buffer
  size on the network from 8kb to 64kb.

Fixed: 83944
Change-Id: I511122b7f35263f05463c84cbfcb8be8f5c60694
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/594994
Commit-Queue: Amit Uttamchandani <amituttam@google.com>
Reviewed-by: Andy Weiss <dragonbear@google.com>