[build] Add package_manifest to the test spec
Infrastructure currently sends all test packages to all test shards,
even though each shard only executes a subset of these tests. We'd like
to change this, and only distributed the necessary packages to each
shard. This requires that each test knows the blobs it needs; this CL
accomplishes this by adding the path to the package manifest (relative
to the root build directory) to the test spec.

Bug: b/195343539
Change-Id: Ia6420cfeee5a9fc45acfff8edc28958318304527
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/596741
Reviewed-by: Anthony Fandrianto <atyfto@google.com>
Reviewed-by: Oliver Newman <olivernewman@google.com>
Commit-Queue: Anirudh Mathukumilli <rudymathu@google.com>