[bt][mock-piconet-server] Clean up request handling
Background: The main task of the server selects over a number of futures
while handling FIDL requests. Upon reception of a MockPeer request, it
invokes a handler which asynchronously sends a message. The receiving
end of the message is also processed by the aforementioned main task,
and handling of said request is blocked on sending the message. Our
theory is that this could have resulted in a sort of circular
dependency, causing flakes due to stalls.

Instead of asynchronously sending the result of the handler, we can
simply return the result synchronously.

While here, improve logs.

Bug: 82484
Test: fx test mock-piconet-server-tests
Test: fx test bt-rfcomm-integration-tests (which rely on the Mock
Piconet Server)

Multiply: bt-rfcomm-integration-tests

Change-Id: I2a4be2f00871ebbbca5cda54e02abbe5d9139691
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/593056
Commit-Queue: Ani Ramakrishnan <aniramakri@google.com>
Fuchsia-Auto-Submit: Ani Ramakrishnan <aniramakri@google.com>
Reviewed-by: Lucas Jenkins <lucasjenkins@google.com>