[workstation] Attempt to get an account directory on launch.
With the current state of the backend implementation, this reaches the
call to get account with fails with UNSUPPORTED_OPERATION, this (and any
other account-related failures) is logged but is not fatal.

I reason that there is little value in adding unit testing to
workstation session for this feature we will be moving into login shell
within a few weeks, but I can do that if necessary.

Bug: 81601
Change-Id: I84b672c4dfeb3f6e2179429b83720b92cbfcf827
Reviewed-on: https://fuchsia-review.googlesource.com/c/experiences/+/596761
Reviewed-by: Sanjay Chouksey <sanjayc@google.com>
Commit-Queue: Jody Sankey <jsankey@google.com>