[ffx] Add portmap to device_spec
As part of our project defining a device specification for `ffx emu`, we
want to specify the base set of ports that should be mapped between the
host and guest systems.

This CL introduces a port_map entry to the device_spec data structure.
It contains fields for the host and guest addresses and port numbers,
the IP version, and the protocol (tcp or udp). It includes a to_string()
function which converts the values into a valid QEMU portmap string, so
we can send it straight to the emulator (or device_launcher).

At this time, if a device_spec contains a port_map, the launcher will
populate the emulator command line with the specified ports. It will
also check for a map to port 22, and add one if it's not included.
However, if the user includes a port_map on the command line, that will
completely override the map in the device_spec (this will be adjusted in
the next CL such that the two are reconciled instead). If neither is
specified, the launcher will add a map to port 22 as always.

The unit tests have been adjusted to continue working, but are not
testing the new functionality. This will be included in the
reconciliation CL, since we don't want to write tests twice.

Change-Id: Ic11e0248539c5a548b10d26cf81b03ceafcbdca3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/596703
Reviewed-by: Yuan Zhi <yuanzhi@google.com>
Commit-Queue: Daniel Dougherty <doughertyda@google.com>