[inspector] Support for GWP-ASan exceptions
inspector_print_debug_info will now print the GWP-ASan information if
the thread is on exception and the exception is caused by GWP-ASan. The
information contains the error type and the allocation stack, as well as
the free stack if the allocation is freed.

Test: in a Fuchsia shell, `export SCUDO_OPTIONS=GWP_ASAN_Enabled=true:\
      GWP_ASAN_SampleRate=1:GWP_ASAN_MaxSimultaneousAllocations=512` and
      run `crasher use_after_free`. The use-after-free is captured by
      GWP-ASan and the crashsvc prints extra info.

Fixed: 77557
Change-Id: I84fd1b8df5ed23f291eb558e1708be1be1ed7a60
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/595642
Reviewed-by: Shai Barack <shayba@google.com>
Reviewed-by: Francois Rousseau <frousseau@google.com>
Commit-Queue: Dangyi Liu <dangyi@google.com>