[media/vnext] add the rest of vnext/lib/stream_sink
This CL adds the remainder of the stream_sink lib, stream_queue
having been added previously. Included are a StreamSink client and a
StreamSink service implementation. The former pulls from a
StreamQueue while the latter pushes to one. This pipeline effectively
proxies a stream queue across a process boundary.

In addition to the client a service implementation, this CL includes
a few files that define associated mechanisms, including a fidl-
specific clear request, a fidl-specific release fence for packets,
and a template for conversion between fidl and internal packet
representations. These are tested along with the client and
service implementation in the unit tests.

Test: new unit tests
Change-Id: I2eddf9cd16169ac13e7bb7a7e8837c0179fa05d7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/597182
Commit-Queue: Dale Sather <dalesat@google.com>
Reviewed-by: Tom Bergan <tombergan@google.com>
Reviewed-by: Austin Foxley <afoxley@google.com>