[kernel][spinlock] Check that ints are disabled in IsHeld().
It is invalid to call SpinLock::IsHeld() with interrupts enabled,
due to the use of the CPU number to indicate the holder.

Add a debug assert that interrupts are disabled when calling
IsHeld() to avoid recurring mistakes using this method in the
wrong state.

- Add debug assert for ints disabled to IsHeld().
- Simplify code that asserts ints are disabled + IsHeld().
- Fix call to IsHeld() that may occur without ints disabled.

Test: Updated test does not assert.
Change-Id: Iee067ceb4c1f158436a4511edb6463379104f148
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/460772
Reviewed-by: Nick Maniscalco <maniscalco@google.com>
Reviewed-by: Travis Geiselbrecht <travisg@google.com>
Reviewed-by: John Grossman <johngro@google.com>
Commit-Queue: Corey Tabaka <eieio@google.com>