[audio][hda] Check for revision only if we are not in reset
The HDA spec states that the driver must read a 1 from the global
control register HwInit bit before accessing any controller registers.
This CL moves the version controller registers reads to after
HwInit is 1.

If we are not held in reset as we boot, we access registers that
could be dependent on version, hence we check for the version first.
After we reset the HW and HwInit goes back to 1, we also check for
the version for cases when we are held in reset as we boot and hence
did not check for version earlier.

This CL re-enables device-enumeration test checks for the HDA driver
as the flake reported in fxbug.dev/84037 is expected to be fixed by
this CL.

Fixed: 84037

Test: device-enumeration-test on NUC

Change-Id: I1d7fe43b3e0f6f725814302751a12445bbecf6fd
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/597043
Commit-Queue: Andres Oportus <andresoportus@google.com>
Reviewed-by: David Greenaway <dgreenaway@google.com>