[ffx] Remove spurious test from common/lib
During a recent refactor, an existence test was added to debug build
issues. It was left in the module when the refactor was complete, but it
adds no value to the test suite. This CL removes it.

Change-Id: I5bf9ad7c1370889c27367c6fd918794212531f09
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/598062
Fuchsia-Auto-Submit: Daniel Dougherty <doughertyda@google.com>
Reviewed-by: Yuan Zhi <yuanzhi@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>