[factoryfs] Update to use new Query interface.
There is a new QueryFilesystem interface in the VFS layer that
simplifies some of the implementation. Use this. The custom QueryService
in factoryfs can now also be replaced with the VFS implementation.

To enable this, the VFS pointer needed to be plumbed into the Factoryfs
class which meant updating the constructors. As part of this, the
::Create functions now use the new zx::status syntax. The initialization
now matches Blobfs again which is what this code was originally based
upon.

Remove an unused function in Blobfs I noticed.

Change-Id: Ic927fe5a2bdce157ec8892fa6adcaf100fb1df66
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/598114
Reviewed-by: Brandon Castellano <bcastell@google.com>
Commit-Queue: Brett Wilson <brettw@google.com>