[starnix] query_events
In FileOps::wait_async, we check the current events
and immediately queue a packet if the the events that
we are interested in are present. We need to do this
so that waiting on Zircon and non-Zircon file objects
have the same behavior. All Zircon kernel waits behave
this way.

The achieve this, we need query_events to check the
current event state of the FileObject.

Bug: 86403

Change-Id: Ib389f7e75c7722f47e6aed4cf384607942b58043
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/594965
Reviewed-by: Adam Barth <abarth@google.com>
Commit-Queue: Steve Austin <steveaustin@google.com>