[bt][a2dp] Fix leaks in Permits, cleanup
With revokable permits, a revokation function could hold a
reference to the Permits. This would cause a reference loop to the
PermmitsInner when dropping the permits, meaning it would leak if
everything was dropped at once.

Have the Permit and the Reservation hold the Revokation function
themselves, so if they are dropped, they will drop their own references
to the Permits.

Some other cleanup as well:
 - Cleanup unused results lints and turn on linter.
 - Add some logs on unexpected results.
 - Switch from log to tracing.
 - Switch to fuchsia::test where reasonable for logs in tests.

Bug: 75375, 86726, 81181
Test: updated, fx test bt-a2dp-lib-tests
Test: run LSAN and confirm no leaks
Change-Id: Ie8413a5d2a4d5209b3d9505a5ded338ae8416bd1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/569544
Commit-Queue: Marie Janssen 💖 <jamuraa@google.com>
Reviewed-by: Ani Ramakrishnan <aniramakri@google.com>