[media/vnext] add vnext/lib/threads
'threads' is a small library that implements the fmlib::Thread class.
fmlib::Thread is essentially a wrapper for a shared pointer to an
async loop and executor. This provides a way to assign one or more
'machines' (e.g. the processor in a software decoder, an input
connector, etc) to a particular thread, automatically destroying
the thread when all assigned machines are destroyed.

A wrapper is used instead of a raw std::shared_ptr to eliminate the
concern that such a pointer might be null (fmlib::Thread always
contains a valid pointer to a loop/executor struct) and to provide
better ergonomics.

Test: new unit test
Change-Id: I356f9a3c5da7bb25373f5fa4c3a9387951ce5fd8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/598282
Reviewed-by: Tom Bergan <tombergan@google.com>
Commit-Queue: Dale Sather <dalesat@google.com>