[vfs] Vnode::Serve take raw zx::channel
In fuchsia.io v1 the `Open` call takes a server end of
`fuchsia.io/Node`, and as such, Vnode::Serve also accepted a
`fuchsia.io/Node`. However, this was not entirely accurate: it's
possible to serve a custom FIDL protocol, as in the case of
`fs::Service`.

fuchsia.io v2 in fact defines the server end during Open to just
a raw channel. This CL updates the Vnode::Serve to align with that as
well, which simplifies a lot of call sites.

This was mainly a yak shaving when migrating vfs to use typed LLCPP
channels.

Change-Id: I3c394d836fbeff61a36dc36e57a7856b3836eb75
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/597825
Reviewed-by: John Bauman <jbauman@google.com>
Reviewed-by: Chris Suter <csuter@google.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Reviewed-by: Austin Foxley <afoxley@google.com>
Fuchsia-Auto-Submit: Yifei Teng <yifeit@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>