[modular][flake] Deflake session/agent restart tests
The is_running check is not consistent, likely because the same
FakeSessionShell instance is used when the component instance is
torn down and recreated.

Track the total number of times the shell component starts up
so the test does not rely on the timing of is_running being set.

Fixed: 87344
Fixed: 87293
Change-Id: I840eb608e8c5bba86febeca37383bc9361e05a1a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/598341
Reviewed-by: Gabe Schine <thatguy@google.com>
Commit-Queue: Yegor Pomortsev <ypomortsev@google.com>