[ffx] Add more info to daemon link failures
There was a failure in the e2e tests that pointed to a failure with the
ascendd link with what look likes a missing file for the socket. This
adds some extra info to the error messages so that if this happens again
it will be easier to figure out what is going on.

Change-Id: If49f1c9081330813cdbf69e239d4add207531af9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/588330
Reviewed-by: Andrew Davies <awdavies@google.com>
Reviewed-by: Jordon Wing <jwing@google.com>
Reviewed-by: James Tucker <raggi@google.com>
Commit-Queue: Andy Weiss <dragonbear@google.com>