[password_authenticator] Abstract disk ops behind DiskManager trait
Instead of AccountManager being parameterized by a `PartitionManager`
trait, the entire disk management layer is represented by a
`DiskManager` trait.

This allows for the mocks in AccountManager to be substantially less
opinionated about the actual bits on the disk.

Test coverage of `DevDiskManager` is more complete because the mocking
is done at the FIDL layer, which guarantees that the client side of
`DevDiskManager` is fully tested.

Bug: 86859
Change-Id: I3e027cb1a18d3d6a809c208d8e31c96daa456845
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/598302
Commit-Queue: Adam Lesinski <adamlesinski@google.com>
Reviewed-by: Jody Sankey <jsankey@google.com>
Reviewed-by: Drew Fisher <zarvox@google.com>