[netdevice-migration] Add basic GetInfo impl
Add a basic implementation of NetworkDeviceImplGetInfo() returning
values appropriate for the generic ethernet driver for which
NetdeviceMigration is a shim layer. The return value can be defined as
a constant without consulting EthernetImplQuery because the latter's
fields are not relevant to device_info_t fields.

Fixed: 80613

Change-Id: I2155d949bf3be78966c0a5912de3ec7fa588b48d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/598442
Commit-Queue: Amanda Tait <atait@google.com>
Reviewed-by: Tamir Duberstein <tamird@google.com>
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>