[ffx] Add example device_spec manifest file
This is an example device_spec, a JSON file which contains emulator
configuration values. The spec is still in active development, so it is
likely to change over time. We plan to follow up with a test which
consumes this test file and generates command lines according to its
contents, but this will be in a later CL.

Change-Id: I0bb551393fe00e628cf13c5765611f02d452f434
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/598741
Fuchsia-Auto-Submit: Daniel Dougherty <doughertyda@google.com>
Reviewed-by: Clayton Wilkinson <wilkinsonclay@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>