[core-tests] Fix flaky test for CPU perf scale bandwidth.
Change the SystemCpu.ScaleBandwidth test to ignore runtime in excess of
the expected value, which may happen in emulated environments on shared
host infra. This is valid because the test is meant to detect receiving
less than the expected runtime when the CPU performance scale is < 1.0.

Fixed: 85846
Change-Id: Ifdf0ff30b3aa0c235602e947180c9ea4899e610b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/597052
Reviewed-by: Nick Maniscalco <maniscalco@google.com>
Commit-Queue: Corey Tabaka <eieio@google.com>