[driver_runtime] Add per-thread driver call stack tracking.
This will be used to implement re-entrancy checking in the
dispatcher. We need to know which driver owns a dispatcher,
so every time we enter a driver we will push it onto the
call stack, and check the top of the stack when
fdf_dispatcher_create() is called.

There is a new internal.h API included in the runtime shared
library that is restricted to the driver host. This is so it
can provide the runtime driver updates on driver entries, as
driver host threads are separate to runtime threads which
the runtime has no access to. e,g. e.g. the client could create
a fdf_dispatcher in a bind call.

Bug: 85577

Change-Id: I579ae503381dde16416e84499b3c828d239bcbdb
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/596062
Commit-Queue: Jocelyn Dang <jocelyndang@google.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>