[component_hub] Support serde on host
We want to expose some of these types, namely list::Component and
show::Component, to be serialized by serde for machine consumable output
inside an ffx plugin. This adds Serialize/Deserialize to these types and
gates that to the host compiled version of this library.

Change-Id: I398b021c04cb633f22b3aa54321949a5e9300a6e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/591901
Commit-Queue: Andy Weiss <dragonbear@google.com>
Reviewed-by: Xyan Bhatnagar <xbhatnag@google.com>