[ffx] Make trace list-categories better
There is a FIDL API called GetKnownCategories which returns a list of
known tracing categories. This is a hard coded list which is in a
config.json file in the trace manager code. It is a subset of categories
which may be available for tracing.

Separately, ffx trace by default enables a list of categories which
intersects the list described above but neither is a subset of the
other. Therefore what should `ffx trace list-categories` show?

This CL adds the default categories to the output. In the human readable
format, I have `Known Categories:` followed by the list returned from
the FIDL call, then `Default Categories:` followed by the constant list
that is defined in the plugin. In the machine output we add known and
default bools to describe where that category comes from. This lets you
have maximal flexibility as a consumer and you can decide what you want
to do with that information.

Also there were tests in the args.rs but they were never run because it
wasn't specified in the BUILD file. This makes those run and adds some
more.

Bug: 86495
Change-Id: Ie6a0d2184cf6879f3902aff3b42adc36b41a04f0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/595146
Reviewed-by: Matthew Boetger <boetger@google.com>
Commit-Queue: Andy Weiss <dragonbear@google.com>