[wlanif] Only set AP online after start confirmed
An AP interface had its ethernet status set to online as soon as the
request to start the AP came in. Additionally it would also remain
online even if the start confirmation indicated that the start failed.

This change makes it so that an AP interface is only considered online
once a successful start confirmation is received. On a failed start
confirmation the interface is set to offline.

Test: fx test wlanif_unittest
Bug: 86875
Change-Id: Iae46d336630655305f137517cf977276089c639d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/595447
Reviewed-by: Joshua Conner <joshconner@google.com>
Commit-Queue: Bjoern Johansson <bjoernj@google.com>