[password_authenticator] Provision a new enrypted account
Implements account provisioning by finding the account data partition
and binding the zxcrypt driver to that block device.

zxcrypt is then used to format the block with a key of zeroes in the
keyslot 0.

The last step in provisioning is to format minfs on the encrypted block
device, which is not implemented in this change.

Bug: 86859
Change-Id: I3763acb82cfd41d2c95d30fd9e30cc8c32378cb9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/597832
Reviewed-by: Jody Sankey <jsankey@google.com>
Commit-Queue: Adam Lesinski <adamlesinski@google.com>