[wlanif] Support notifying driver of online status
Some drivers, especially those with separate data planes, might need to
know about when the online status of an interface changes. Currently the
wlanif driver handles this and the underlying driver does not know
anything about this.

Add a method to the FIDL interface to allow wlanif to notify the driver
if it's interested.

Test: fx test wlanif_unittests
      build bringup.x64 with drivers-only
      build bringup.arm64 with drivers-only
Change-Id: I2f9d5b6447070578facee715b9a110e4692f3c96
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/593262
Commit-Queue: Bjoern Johansson <bjoernj@google.com>
Reviewed-by: Sakthi Vignesh Radhakrishnan <rsakthi@google.com>
Reviewed-by: Joshua Conner <joshconner@google.com>
Reviewed-by: Rebecca Silberstein <silberst@google.com>
Reviewed-by: Charles Celerier <chcl@google.com>
API-Review: Rebecca Silberstein <silberst@google.com>