[ffx] move daemonize tests to a standalone crate
We started to observe a deadlock from this test on Linux, which is
likely due to an improperly handled failure to spawn the child process,
resultant from CLOEXEC behavior (exact root cause was not identified).

Moving this test to it's own crate ensures that the daemonize call
occurs at an appropriately early and isolated place in a program
runtime, comparable with the intended use case of the daemonize()
function.

Change-Id: I83048d6ead62d883629974cc28384fd4610125ad
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/599525
Commit-Queue: James Tucker <raggi@google.com>
Fuchsia-Auto-Submit: James Tucker <raggi@google.com>
Reviewed-by: Andrew Davies <awdavies@google.com>