[wlan][policy] Detect multiple controllers.
The existing way to check if another service holds a policy
controller is to check if the channel has been closed.  This
works on Fuchsia, but is not an option on host tools that
make use of overnet.  This change introduces a wrapper to
test the results of policy commands to check for the epitaph
that the policy layer sends.

The following tests were run to ensure that the error was is
printed when another service holds the policy controller:

Test: donut client scan
Test: donut ap stop --ssid asdf
Test: fx test donut-tests
Multiply: donut-tests
Bug: 87482
Change-Id: Ib52c6c2f1a1e01c9642e904b47a4cc96828ca368
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/598403
Commit-Queue: Kevin Sakuma <sakuma@google.com>
Reviewed-by: Marc Khouri <mnck@google.com>