[build] Delete dead condition for zircon optimize level
The body of the condition includes a reference to a target that's been gone for a long time.
Apparently this is never exercised because if it were then it'd fail to gen.
Not sure the underlying problem exists anymore?

Change-Id: I55beb008d98a6c880a8ebcc3aaa568616820f87f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/594621
Fuchsia-Auto-Submit: Shai Barack <shayba@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Jay Zhuang <jayzhuang@google.com>