[iomisc] Make ReadUntilMatch easier to call
Have it take a reader and targets to match and construct a
MatchingReader itself, instead of requiring the caller to construct and
pass in a MatchingReader.

Also add a new ReadUntilMatchString wrapper function for the common case
where we want to match a string, and it's annoying for the caller to
need to convert the string to a byte slice.

Change-Id: Ib5cb1336b6aeb12db96e09baf45f045f917b5173
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/598701
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Oliver Newman <olivernewman@google.com>
Reviewed-by: Anirudh Mathukumilli <rudymathu@google.com>
Reviewed-by: Ina Huh <ihuh@google.com>