[intel][wlan] Make all devices have the same parent
Currently the iwlwifi creates a root device
that is bindable then adds children to that
device.  This is a behavior that is incompatible
with the next iteration of the driver framework.

As a workaround to avoid this behavior, this CL
causes the iwlwifi-wlanphyimpl to add children to
its parent instead of itself.

Bug: 84362
Test: device enumeration test, manual testing.
Change-Id: Icceb28ec32d5d3489c7048728148306f96c4ad26
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/596702
Commit-Queue: Garratt Gallagher <garratt@google.com>
Reviewed-by: Louis Yung-chieh Lo <yjlou@google.com>