[ftl] Add diagnosis output to volume extractor.
Adds a diagnosis section for analyzing known issues in an FTL. This
makes it easy to reach inside the FTL and check on various conditions to
identify known symptoms of issues that have been found to corrupt an FTL
volume. Including a check for Partial Page Writes, as well as
specifically checking for them when the fix has already been applied.
Also looks for vpage mapping collisions to identify Premature Block
Recycle.

A couple of problems have just had their logging adjusted to make
spotting the issue more clear, such as the failed read during load time,
and also a vpage mapping that points to a non volume block which would
indicate Premature Block Recycle.

Also cleans up some of the logging output, like iterating the correct
number of vpages when dumping to remove the harmless error log, and no
longer logging when bad blocks are found, as that is a normal and safe
occurrence.

Tested: Against 4 different images with known problems in them, and 1
image that was doctored to induce PPW with the fix.

Change-Id: I43855c1d0e414db0398fab712da19eef8291bcff
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/599265
Commit-Queue: Martin Lindsay <mlindsay@google.com>
Reviewed-by: James Sullivan <jfsulliv@google.com>