[ffx] Fix flake in ffx_emulator_common_lib_test
The test for download_and_extract has been flaking for some time,
because it actually calls out to CIPD and downloads and extracts a file
multiple times on each invocation. To stop this, this CL mocks out the
actual downloading and extraction, and just tests the code in HostTools
which coordinates the calls to CIPD.

We use the mockall and async_trait crates to define a new trait and
implementing struct around the existing CIPD code, then mock the CIPD
functions for the test. The mocks never make http calls, or extract any
files; rather we just ensure the calls themselves are made, and that
the HostTools code reacts correctly according to the return values of
those calls. By removing the test's dependency on CIPD and the network,
we expect this will resolve the flakiness entirely.

Bug: 85945
Change-Id: I55c4e93b263e9b7f1e1142cc4c65052e54202632
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/599074
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Dave Schuyler <dschuyler@google.com>
Fuchsia-Auto-Submit: Daniel Dougherty <doughertyda@google.com>