[ffx][emu] Making graphic_utils VS code friendly
This adds a function that is available for all configurations which
wraps the os specific version of these calls.

This fixes the issue with rust-analyzer not being able to correctly
parse the file resulting in a red sqiggle underline.

Change-Id: I81d2166efc480fcbf14bbe954c570db0e8967be1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/600060
Fuchsia-Auto-Submit: Clayton Wilkinson <wilkinsonclay@google.com>
Reviewed-by: Daniel Dougherty <doughertyda@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>